requirements.txt
psycopg2-binary==2.9.9
numpy==1.24.3
scipy==1.11.4
